Support Int32 parsing by default; add some more dev docs

+++ test/basics.jl
@@ -308,7 +308,7 @@ f = CSV.File(IOBuffer("int,float,date,datetime,bool,null,str,catg,int_float\n1,3
 @test Tables.istable(f)
 @test Tables.rowaccess(typeof(f))
 @test Tables.columnaccess(typeof(f))
-@test Tables.schema(f) == Tables.Schema([:int, :float, :date, :datetime, :bool, :null, :str, :catg, :int_float], [Int64, Float64, Date, DateTime, Bool, Missing, InlineString7, InlineString3, Float64])
+@test Tables.schema(f) == Tables.Schema([:int, :float, :date, :datetime, :bool, :null, :str, :catg, :int_float], [Int32, Float64, Date, DateTime, Bool, Missing, InlineString7, InlineString3, Float64])
 @test Tables.rows(f) === f
 @test eltype(f) <: CSV.Row
 row = first(f)
@@ -426,7 +426,7 @@ f = CSV.File(transcode(GzipDecompressor, Mmap.mmap(joinpath(dir, "randoms.csv.gz
 @test f.wage isa AbstractVector{Union{Missing, Dec64}}
 
 f = CSV.File(joinpath(dir, "promotions.csv"); stringtype=PosLenString)
-@test Tables.schema(f).types == (Float64, Union{Missing, Int64}, Union{Missing, Float64}, PosLenString, Union{Missing, PosLenString}, PosLenString, PosLenString, Union{Missing, Int64})
+@test Tables.schema(f).types == (Float64, Union{Missing, Int32}, Union{Missing, Float64}, PosLenString, Union{Missing, PosLenString}, PosLenString, PosLenString, Union{Missing, Int32})
 
 f = CSV.File(joinpath(dir, "promotions.csv"); limit=7500, threaded=true)
 @test length(f) == 7500
@@ -437,7 +437,7 @@ f = CSV.File(IOBuffer("1,2\r\n3,4\r\n\r\n5,6\r\n"); header=["col1", "col2"], ign
 f = CSV.File(joinpath(dir, "escape_row_starts.csv"); tasks=2)
 @test length(f) == 10000
 @test eltype(f.col1) == InlineString63
-@test eltype(f.col2) == Int64
+@test eltype(f.col2) == Int32
 
 f = CSV.File(IOBuffer("col1\nhey\nthere\nsailor"); stringtype=PosLenString)
 @test f.col1 isa PosLenStringVector
@@ -487,9 +487,9 @@ f = CSV.File(buf, header=["A", "B"])
 @test (f[2].A, f[2].B) == ("1", "2")
 
 # 680: ensure typemap works with custom types
-f = CSV.File(IOBuffer("a\n1\n2\n3"); typemap=Dict(Int64=>Int32))
+f = CSV.File(IOBuffer("a\n1\n2\n3"); typemap=Dict(Int32=>Int32))
 @test f.a isa Vector{Int32}
-f = CSV.File(IOBuffer("a\n1\n2\n3"); typemap=Dict(Int64=>String))
+f = CSV.File(IOBuffer("a\n1\n2\n3"); typemap=Dict(Int32=>String))
 @test f.a isa Vector{String}
 
 # 678
